memex/requirements.txt
# Core LLM client
openai>=1.30.0

# API framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Data / async HTTP
httpx>=0.27.0
pydantic>=2.7.0

# Reddit
praw>=7.7.1

# Database
asyncpg>=0.29.0
pgvector>=0.2.5     # psycopg2 adapter for pgvector types

# Report rendering
markdown>=3.6
weasyprint>=62.0    # optional: PDF export

# Local embedding model (ingest pipeline relevance scoring)
# Set NEWS_USE_LOCAL_EMBEDDINGS=false to skip this dependency
sentence-transformers>=3.0.0

# Utilities
python-dotenv>=1.0.0
